removed WarehosueItem, too much completxity
